stdlib: Move default exit handler behavior to exit_handler.py
Instead of specifying all of this in the simulator, move it to the class
that implements the classic behvaior.
